Fix variable type in Stair model

+++ Stair/config.h
@@ -20,7 +20,7 @@ typedef enum {
 
 typedef struct {
 
-    double counter;
+    int counter;
 
 } ModelData;
